Move sanitizeDecimalAmount to utils.ts
